+++ src/egon/data/datasets/vg250/__init__.py
@@ -1,4 +1,4 @@
-"""The central module containing all code dealing with importing VG250 data.
+"""The central module containing all code dealing with VG250 data.
 
 This module either directly contains the code dealing with importing VG250
 data, or it re-exports everything needed to handle it. Please refrain
@@ -12,6 +12,7 @@ isn't exported from this module, please file a bug, so we can fix this.
 from urllib.request import urlretrieve
 import json
 import os
+import codecs
 
 from geoalchemy2 import Geometry
 import geopandas as gpd
@@ -19,9 +20,10 @@ import geopandas as gpd
 from egon.data import db
 from egon.data.config import settings
 import egon.data.config
+from egon.data.datasets import Dataset
 
 
-def download_vg250_files():
+def download_files():
     """Download VG250 (Verwaltungsgebiete) shape files."""
     data_config = egon.data.config.datasets()
     vg250_config = data_config["vg250"]["original_data"]
@@ -270,4 +272,49 @@ def add_metadata():
 
         db.submit_comment(
             meta_json, vg250_config["processed"]["schema"], table
+        )
+
+def run_sql_script(script):
+    """ Runs SQL script in egon-data database
+
+    Parameters
+    ----------
+    script : str
+        Name of SQL-script
+
+    Returns
+    -------
+    None.
+
+    """
+
+    with codecs.open(script, "r", "utf-8-sig") as fd:
+        sqlfile = fd.read()
+
+    db.execute_sql(sqlfile)
+
+
+def nuts_mview():
+
+    run_sql_script(os.path.join(os.path.dirname(__file__),
+                                'vg250_lan_nuts_id_mview.sql'))
+
+
+def cleaning_and_preperation():
+
+    run_sql_script(os.path.join(os.path.dirname(__file__),
+                                'cleaning_and_preparation.sql'))
+
+class Vg250(Dataset):
+
+    filename = egon.data.config.datasets(
+        )["vg250"]["original_data"]["source"]["url"]
+
+    def __init__(self, dependencies):
+        super().__init__(
+            name="VG250",
+            version=self.filename +"-0.0.0",
+            dependencies=dependencies,
+            tasks=(download_files, to_postgres, nuts_mview,
+                   add_metadata, cleaning_and_preperation),
         )